Splitted nodeDefDetails and nodeDefDetailsAdvanced helper functions
