+++ test/e2e/tests/013-surveyFormEditExpressions.js
@@ -1,13 +1,13 @@
 import { click } from '../utils/api'
 import { editNodeDef } from '../utils/ui/surveyForm'
 import { clickSidebarBtnSurveyForm } from '../utils/ui/sidebar'
+import { clickNodeDefSaveAndBack } from '../utils/ui/nodeDefDetail'
 import {
   addNodeDefDefaultValue,
-  clickNodeDefSaveAndBack,
   expectNodeDefDefaultValue,
   expectNodeDefRelevantIf,
   setNodeDefRelevantIf,
-} from '../utils/ui/nodeDefDetail'
+} from '../utils/ui/nodeDefDetailsAdvanced'
 
 describe('SurveyForm edit expressions', () => {
   test('add Default Value to "Cluster decimal"', async () => {
